'Apply changes and move to shared folder'
