+++ src/shared/components/ui/SkeletonCard.tsx
