--- src/stores/useAuthStore.ts
@@ -1,11 +1,0 @@
-import { create } from "zustand";
-
-interface AuthState {
-  user: any; 
-  setUser: (user: any) => void;
-}
-
-export const useAuthStore = create<AuthState>((set) => ({
-  user: null,
-  setUser: (user) => set({ user }),
-}));
